requirements.txt
streamlit
google-cloud-bigquery
google-cloud-bigquery[pandas]==2.34.3
pydeck